% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_details.R
\docType{class}
\name{JobDetails}
\alias{JobDetails}
\title{JobDetails}
\format{
An \code{R6Class} generator object
}
\description{
JobDetails Class

JobDetails Class
}
\details{
Create a new JobDetails
}
\examples{

## ------------------------------------------------
## Method `JobDetails$toList`
## ------------------------------------------------

# convert array of JobDetails (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aktuelleVeroeffentlichungsdatum}}{character [optional]}

\item{\code{angebotsart}}{character [optional]}

\item{\code{arbeitgeber}}{character [optional]}

\item{\code{branchengruppe}}{character [optional]}

\item{\code{branche}}{character [optional]}

\item{\code{arbeitgeberHashId}}{character [optional]}

\item{\code{arbeitsorte}}{list(\link{JobDetailsArbeitsorteInner}) [optional]}

\item{\code{arbeitszeitmodelle}}{list(character) [optional]}

\item{\code{befristung}}{character [optional]}

\item{\code{uebernahme}}{character [optional]}

\item{\code{betriebsgroesse}}{character [optional]}

\item{\code{eintrittsdatum}}{character [optional]}

\item{\code{ersteVeroeffentlichungsdatum}}{character [optional]}

\item{\code{allianzpartner}}{character [optional]}

\item{\code{allianzpartnerUrl}}{character [optional]}

\item{\code{titel}}{character [optional]}

\item{\code{hashId}}{character [optional]}

\item{\code{beruf}}{character [optional]}

\item{\code{modifikationsTimestamp}}{character [optional]}

\item{\code{stellenbeschreibung}}{character [optional]}

\item{\code{refnr}}{character [optional]}

\item{\code{fuerFluechtlingeGeeignet}}{character [optional]}

\item{\code{nurFuerSchwerbehinderte}}{character [optional]}

\item{\code{anzahlOffeneStellen}}{integer [optional]}

\item{\code{arbeitgeberAdresse}}{\link{JobDetailsArbeitgeberAdresse} [optional]}

\item{\code{fertigkeiten}}{list(\link{JobDetailsFertigkeitenInner}) [optional]}

\item{\code{mobilitaet}}{\link{JobDetailsMobilitaet} [optional]}

\item{\code{fuehrungskompetenzen}}{\link{JobDetailsFuehrungskompetenzen} [optional]}

\item{\code{verguetung}}{character [optional]}

\item{\code{arbeitgeberdarstellungUrl}}{character [optional]}

\item{\code{arbeitgeberdarstellung}}{character [optional]}

\item{\code{hauptDkz}}{character [optional]}

\item{\code{istBetreut}}{character [optional]}

\item{\code{istGoogleJobsRelevant}}{character [optional]}

\item{\code{anzeigeAnonym}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobDetails-new}{\code{JobDetails$new()}}
\item \href{#method-JobDetails-toJSON}{\code{JobDetails$toJSON()}}
\item \href{#method-JobDetails-toList}{\code{JobDetails$toList()}}
\item \href{#method-JobDetails-toSimpleType}{\code{JobDetails$toSimpleType()}}
\item \href{#method-JobDetails-fromJSON}{\code{JobDetails$fromJSON()}}
\item \href{#method-JobDetails-toJSONString}{\code{JobDetails$toJSONString()}}
\item \href{#method-JobDetails-fromJSONString}{\code{JobDetails$fromJSONString()}}
\item \href{#method-JobDetails-validateJSON}{\code{JobDetails$validateJSON()}}
\item \href{#method-JobDetails-toString}{\code{JobDetails$toString()}}
\item \href{#method-JobDetails-isValid}{\code{JobDetails$isValid()}}
\item \href{#method-JobDetails-getInvalidFields}{\code{JobDetails$getInvalidFields()}}
\item \href{#method-JobDetails-print}{\code{JobDetails$print()}}
\item \href{#method-JobDetails-clone}{\code{JobDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobDetails class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$new(
  aktuelleVeroeffentlichungsdatum = NULL,
  angebotsart = NULL,
  arbeitgeber = NULL,
  branchengruppe = NULL,
  branche = NULL,
  arbeitgeberHashId = NULL,
  arbeitsorte = NULL,
  arbeitszeitmodelle = NULL,
  befristung = NULL,
  uebernahme = NULL,
  betriebsgroesse = NULL,
  eintrittsdatum = NULL,
  ersteVeroeffentlichungsdatum = NULL,
  allianzpartner = NULL,
  allianzpartnerUrl = NULL,
  titel = NULL,
  hashId = NULL,
  beruf = NULL,
  modifikationsTimestamp = NULL,
  stellenbeschreibung = NULL,
  refnr = NULL,
  fuerFluechtlingeGeeignet = NULL,
  nurFuerSchwerbehinderte = NULL,
  anzahlOffeneStellen = NULL,
  arbeitgeberAdresse = NULL,
  fertigkeiten = NULL,
  mobilitaet = NULL,
  fuehrungskompetenzen = NULL,
  verguetung = NULL,
  arbeitgeberdarstellungUrl = NULL,
  arbeitgeberdarstellung = NULL,
  hauptDkz = NULL,
  istBetreut = NULL,
  istGoogleJobsRelevant = NULL,
  anzeigeAnonym = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aktuelleVeroeffentlichungsdatum}}{aktuelleVeroeffentlichungsdatum}

\item{\code{angebotsart}}{angebotsart}

\item{\code{arbeitgeber}}{arbeitgeber}

\item{\code{branchengruppe}}{branchengruppe}

\item{\code{branche}}{branche}

\item{\code{arbeitgeberHashId}}{arbeitgeberHashId}

\item{\code{arbeitsorte}}{arbeitsorte}

\item{\code{arbeitszeitmodelle}}{arbeitszeitmodelle}

\item{\code{befristung}}{befristung}

\item{\code{uebernahme}}{uebernahme}

\item{\code{betriebsgroesse}}{betriebsgroesse}

\item{\code{eintrittsdatum}}{eintrittsdatum}

\item{\code{ersteVeroeffentlichungsdatum}}{ersteVeroeffentlichungsdatum}

\item{\code{allianzpartner}}{allianzpartner}

\item{\code{allianzpartnerUrl}}{allianzpartnerUrl}

\item{\code{titel}}{titel}

\item{\code{hashId}}{hashId}

\item{\code{beruf}}{beruf}

\item{\code{modifikationsTimestamp}}{modifikationsTimestamp}

\item{\code{stellenbeschreibung}}{stellenbeschreibung}

\item{\code{refnr}}{refnr}

\item{\code{fuerFluechtlingeGeeignet}}{fuerFluechtlingeGeeignet}

\item{\code{nurFuerSchwerbehinderte}}{nurFuerSchwerbehinderte}

\item{\code{anzahlOffeneStellen}}{anzahlOffeneStellen}

\item{\code{arbeitgeberAdresse}}{arbeitgeberAdresse}

\item{\code{fertigkeiten}}{fertigkeiten}

\item{\code{mobilitaet}}{mobilitaet}

\item{\code{fuehrungskompetenzen}}{fuehrungskompetenzen}

\item{\code{verguetung}}{verguetung}

\item{\code{arbeitgeberdarstellungUrl}}{arbeitgeberdarstellungUrl}

\item{\code{arbeitgeberdarstellung}}{arbeitgeberdarstellung}

\item{\code{hauptDkz}}{hauptDkz}

\item{\code{istBetreut}}{istBetreut}

\item{\code{istGoogleJobsRelevant}}{istGoogleJobsRelevant}

\item{\code{anzeigeAnonym}}{anzeigeAnonym}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobDetails as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobDetails (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobDetails to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobDetails
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobDetails in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobDetails
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobDetails and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetails-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetails-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
