% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_details_arbeitgeber_adresse.R
\docType{class}
\name{JobDetailsArbeitgeberAdresse}
\alias{JobDetailsArbeitgeberAdresse}
\title{JobDetailsArbeitgeberAdresse}
\format{
An \code{R6Class} generator object
}
\description{
JobDetailsArbeitgeberAdresse Class

JobDetailsArbeitgeberAdresse Class
}
\details{
Create a new JobDetailsArbeitgeberAdresse
}
\examples{

## ------------------------------------------------
## Method `JobDetailsArbeitgeberAdresse$toList`
## ------------------------------------------------

# convert array of JobDetailsArbeitgeberAdresse (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{land}}{character [optional]}

\item{\code{region}}{character [optional]}

\item{\code{plz}}{character [optional]}

\item{\code{ort}}{character [optional]}

\item{\code{strasse}}{character [optional]}

\item{\code{strasseHausnummer}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobDetailsArbeitgeberAdresse-new}{\code{JobDetailsArbeitgeberAdresse$new()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-toJSON}{\code{JobDetailsArbeitgeberAdresse$toJSON()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-toList}{\code{JobDetailsArbeitgeberAdresse$toList()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-toSimpleType}{\code{JobDetailsArbeitgeberAdresse$toSimpleType()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-fromJSON}{\code{JobDetailsArbeitgeberAdresse$fromJSON()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-toJSONString}{\code{JobDetailsArbeitgeberAdresse$toJSONString()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-fromJSONString}{\code{JobDetailsArbeitgeberAdresse$fromJSONString()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-validateJSON}{\code{JobDetailsArbeitgeberAdresse$validateJSON()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-toString}{\code{JobDetailsArbeitgeberAdresse$toString()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-isValid}{\code{JobDetailsArbeitgeberAdresse$isValid()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-getInvalidFields}{\code{JobDetailsArbeitgeberAdresse$getInvalidFields()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-print}{\code{JobDetailsArbeitgeberAdresse$print()}}
\item \href{#method-JobDetailsArbeitgeberAdresse-clone}{\code{JobDetailsArbeitgeberAdresse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobDetailsArbeitgeberAdresse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$new(
  land = NULL,
  region = NULL,
  plz = NULL,
  ort = NULL,
  strasse = NULL,
  strasseHausnummer = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{land}}{land}

\item{\code{region}}{region}

\item{\code{plz}}{plz}

\item{\code{ort}}{ort}

\item{\code{strasse}}{strasse}

\item{\code{strasseHausnummer}}{strasseHausnummer}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobDetailsArbeitgeberAdresse as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobDetailsArbeitgeberAdresse (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobDetailsArbeitgeberAdresse to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobDetailsArbeitgeberAdresse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsArbeitgeberAdresse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobDetailsArbeitgeberAdresse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobDetailsArbeitgeberAdresse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsArbeitgeberAdresse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobDetailsArbeitgeberAdresse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobDetailsArbeitgeberAdresse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitgeberAdresse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitgeberAdresse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitgeberAdresse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
