% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_details_arbeitsorte_inner_koordinaten.R
\docType{class}
\name{JobDetailsArbeitsorteInnerKoordinaten}
\alias{JobDetailsArbeitsorteInnerKoordinaten}
\title{JobDetailsArbeitsorteInnerKoordinaten}
\format{
An \code{R6Class} generator object
}
\description{
JobDetailsArbeitsorteInnerKoordinaten Class

JobDetailsArbeitsorteInnerKoordinaten Class
}
\details{
Create a new JobDetailsArbeitsorteInnerKoordinaten
}
\examples{

## ------------------------------------------------
## Method `JobDetailsArbeitsorteInnerKoordinaten$toList`
## ------------------------------------------------

# convert array of JobDetailsArbeitsorteInnerKoordinaten (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lat}}{numeric [optional]}

\item{\code{lon}}{numeric [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-new}{\code{JobDetailsArbeitsorteInnerKoordinaten$new()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-toJSON}{\code{JobDetailsArbeitsorteInnerKoordinaten$toJSON()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-toList}{\code{JobDetailsArbeitsorteInnerKoordinaten$toList()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-toSimpleType}{\code{JobDetailsArbeitsorteInnerKoordinaten$toSimpleType()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-fromJSON}{\code{JobDetailsArbeitsorteInnerKoordinaten$fromJSON()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-toJSONString}{\code{JobDetailsArbeitsorteInnerKoordinaten$toJSONString()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-fromJSONString}{\code{JobDetailsArbeitsorteInnerKoordinaten$fromJSONString()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-validateJSON}{\code{JobDetailsArbeitsorteInnerKoordinaten$validateJSON()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-toString}{\code{JobDetailsArbeitsorteInnerKoordinaten$toString()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-isValid}{\code{JobDetailsArbeitsorteInnerKoordinaten$isValid()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-getInvalidFields}{\code{JobDetailsArbeitsorteInnerKoordinaten$getInvalidFields()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-print}{\code{JobDetailsArbeitsorteInnerKoordinaten$print()}}
\item \href{#method-JobDetailsArbeitsorteInnerKoordinaten-clone}{\code{JobDetailsArbeitsorteInnerKoordinaten$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobDetailsArbeitsorteInnerKoordinaten class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$new(lat = NULL, lon = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{lat}

\item{\code{lon}}{lon}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobDetailsArbeitsorteInnerKoordinaten as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobDetailsArbeitsorteInnerKoordinaten (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobDetailsArbeitsorteInnerKoordinaten to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobDetailsArbeitsorteInnerKoordinaten
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsArbeitsorteInnerKoordinaten
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobDetailsArbeitsorteInnerKoordinaten in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobDetailsArbeitsorteInnerKoordinaten
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsArbeitsorteInnerKoordinaten
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobDetailsArbeitsorteInnerKoordinaten and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobDetailsArbeitsorteInnerKoordinaten
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsArbeitsorteInnerKoordinaten-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsArbeitsorteInnerKoordinaten-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsArbeitsorteInnerKoordinaten$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
