% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_details_fertigkeiten_inner.R
\docType{class}
\name{JobDetailsFertigkeitenInner}
\alias{JobDetailsFertigkeitenInner}
\title{JobDetailsFertigkeitenInner}
\format{
An \code{R6Class} generator object
}
\description{
JobDetailsFertigkeitenInner Class

JobDetailsFertigkeitenInner Class
}
\details{
Create a new JobDetailsFertigkeitenInner
}
\examples{

## ------------------------------------------------
## Method `JobDetailsFertigkeitenInner$toList`
## ------------------------------------------------

# convert array of JobDetailsFertigkeitenInner (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{hierarchieName}}{character [optional]}

\item{\code{auspraegungen}}{object [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobDetailsFertigkeitenInner-new}{\code{JobDetailsFertigkeitenInner$new()}}
\item \href{#method-JobDetailsFertigkeitenInner-toJSON}{\code{JobDetailsFertigkeitenInner$toJSON()}}
\item \href{#method-JobDetailsFertigkeitenInner-toList}{\code{JobDetailsFertigkeitenInner$toList()}}
\item \href{#method-JobDetailsFertigkeitenInner-toSimpleType}{\code{JobDetailsFertigkeitenInner$toSimpleType()}}
\item \href{#method-JobDetailsFertigkeitenInner-fromJSON}{\code{JobDetailsFertigkeitenInner$fromJSON()}}
\item \href{#method-JobDetailsFertigkeitenInner-toJSONString}{\code{JobDetailsFertigkeitenInner$toJSONString()}}
\item \href{#method-JobDetailsFertigkeitenInner-fromJSONString}{\code{JobDetailsFertigkeitenInner$fromJSONString()}}
\item \href{#method-JobDetailsFertigkeitenInner-validateJSON}{\code{JobDetailsFertigkeitenInner$validateJSON()}}
\item \href{#method-JobDetailsFertigkeitenInner-toString}{\code{JobDetailsFertigkeitenInner$toString()}}
\item \href{#method-JobDetailsFertigkeitenInner-isValid}{\code{JobDetailsFertigkeitenInner$isValid()}}
\item \href{#method-JobDetailsFertigkeitenInner-getInvalidFields}{\code{JobDetailsFertigkeitenInner$getInvalidFields()}}
\item \href{#method-JobDetailsFertigkeitenInner-print}{\code{JobDetailsFertigkeitenInner$print()}}
\item \href{#method-JobDetailsFertigkeitenInner-clone}{\code{JobDetailsFertigkeitenInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobDetailsFertigkeitenInner class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$new(
  hierarchieName = NULL,
  auspraegungen = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hierarchieName}}{hierarchieName}

\item{\code{auspraegungen}}{auspraegungen}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobDetailsFertigkeitenInner as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobDetailsFertigkeitenInner (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobDetailsFertigkeitenInner to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobDetailsFertigkeitenInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsFertigkeitenInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobDetailsFertigkeitenInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobDetailsFertigkeitenInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsFertigkeitenInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobDetailsFertigkeitenInner and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobDetailsFertigkeitenInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFertigkeitenInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFertigkeitenInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFertigkeitenInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
