% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_details_fuehrungskompetenzen.R
\docType{class}
\name{JobDetailsFuehrungskompetenzen}
\alias{JobDetailsFuehrungskompetenzen}
\title{JobDetailsFuehrungskompetenzen}
\format{
An \code{R6Class} generator object
}
\description{
JobDetailsFuehrungskompetenzen Class

JobDetailsFuehrungskompetenzen Class
}
\details{
Create a new JobDetailsFuehrungskompetenzen
}
\examples{

## ------------------------------------------------
## Method `JobDetailsFuehrungskompetenzen$toList`
## ------------------------------------------------

# convert array of JobDetailsFuehrungskompetenzen (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{hatVollmacht}}{character [optional]}

\item{\code{hatBudgetverantwortung}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobDetailsFuehrungskompetenzen-new}{\code{JobDetailsFuehrungskompetenzen$new()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-toJSON}{\code{JobDetailsFuehrungskompetenzen$toJSON()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-toList}{\code{JobDetailsFuehrungskompetenzen$toList()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-toSimpleType}{\code{JobDetailsFuehrungskompetenzen$toSimpleType()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-fromJSON}{\code{JobDetailsFuehrungskompetenzen$fromJSON()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-toJSONString}{\code{JobDetailsFuehrungskompetenzen$toJSONString()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-fromJSONString}{\code{JobDetailsFuehrungskompetenzen$fromJSONString()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-validateJSON}{\code{JobDetailsFuehrungskompetenzen$validateJSON()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-toString}{\code{JobDetailsFuehrungskompetenzen$toString()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-isValid}{\code{JobDetailsFuehrungskompetenzen$isValid()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-getInvalidFields}{\code{JobDetailsFuehrungskompetenzen$getInvalidFields()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-print}{\code{JobDetailsFuehrungskompetenzen$print()}}
\item \href{#method-JobDetailsFuehrungskompetenzen-clone}{\code{JobDetailsFuehrungskompetenzen$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobDetailsFuehrungskompetenzen class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$new(
  hatVollmacht = NULL,
  hatBudgetverantwortung = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hatVollmacht}}{hatVollmacht}

\item{\code{hatBudgetverantwortung}}{hatBudgetverantwortung}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobDetailsFuehrungskompetenzen as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobDetailsFuehrungskompetenzen (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobDetailsFuehrungskompetenzen to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobDetailsFuehrungskompetenzen
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsFuehrungskompetenzen
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobDetailsFuehrungskompetenzen in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobDetailsFuehrungskompetenzen
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsFuehrungskompetenzen
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobDetailsFuehrungskompetenzen and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobDetailsFuehrungskompetenzen
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsFuehrungskompetenzen-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsFuehrungskompetenzen-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsFuehrungskompetenzen$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
