% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_details_mobilitaet.R
\docType{class}
\name{JobDetailsMobilitaet}
\alias{JobDetailsMobilitaet}
\title{JobDetailsMobilitaet}
\format{
An \code{R6Class} generator object
}
\description{
JobDetailsMobilitaet Class

JobDetailsMobilitaet Class
}
\details{
Create a new JobDetailsMobilitaet
}
\examples{

## ------------------------------------------------
## Method `JobDetailsMobilitaet$toList`
## ------------------------------------------------

# convert array of JobDetailsMobilitaet (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reisebereitschaft}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobDetailsMobilitaet-new}{\code{JobDetailsMobilitaet$new()}}
\item \href{#method-JobDetailsMobilitaet-toJSON}{\code{JobDetailsMobilitaet$toJSON()}}
\item \href{#method-JobDetailsMobilitaet-toList}{\code{JobDetailsMobilitaet$toList()}}
\item \href{#method-JobDetailsMobilitaet-toSimpleType}{\code{JobDetailsMobilitaet$toSimpleType()}}
\item \href{#method-JobDetailsMobilitaet-fromJSON}{\code{JobDetailsMobilitaet$fromJSON()}}
\item \href{#method-JobDetailsMobilitaet-toJSONString}{\code{JobDetailsMobilitaet$toJSONString()}}
\item \href{#method-JobDetailsMobilitaet-fromJSONString}{\code{JobDetailsMobilitaet$fromJSONString()}}
\item \href{#method-JobDetailsMobilitaet-validateJSON}{\code{JobDetailsMobilitaet$validateJSON()}}
\item \href{#method-JobDetailsMobilitaet-toString}{\code{JobDetailsMobilitaet$toString()}}
\item \href{#method-JobDetailsMobilitaet-isValid}{\code{JobDetailsMobilitaet$isValid()}}
\item \href{#method-JobDetailsMobilitaet-getInvalidFields}{\code{JobDetailsMobilitaet$getInvalidFields()}}
\item \href{#method-JobDetailsMobilitaet-print}{\code{JobDetailsMobilitaet$print()}}
\item \href{#method-JobDetailsMobilitaet-clone}{\code{JobDetailsMobilitaet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobDetailsMobilitaet class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$new(reisebereitschaft = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reisebereitschaft}}{reisebereitschaft}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobDetailsMobilitaet as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobDetailsMobilitaet (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobDetailsMobilitaet to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobDetailsMobilitaet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsMobilitaet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobDetailsMobilitaet in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobDetailsMobilitaet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobDetailsMobilitaet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobDetailsMobilitaet and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobDetailsMobilitaet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobDetailsMobilitaet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobDetailsMobilitaet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobDetailsMobilitaet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
