% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_search_response_facetten_inner.R
\docType{class}
\name{JobSearchResponseFacettenInner}
\alias{JobSearchResponseFacettenInner}
\title{JobSearchResponseFacettenInner}
\format{
An \code{R6Class} generator object
}
\description{
JobSearchResponseFacettenInner Class

JobSearchResponseFacettenInner Class
}
\details{
Create a new JobSearchResponseFacettenInner
}
\examples{

## ------------------------------------------------
## Method `JobSearchResponseFacettenInner$toList`
## ------------------------------------------------

# convert array of JobSearchResponseFacettenInner (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{befristung}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{behinderung}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{pav}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{berufsfeld}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{arbeitsort}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{ausbildungsart}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{veroeffentlichtseit}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{schulbildung}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{arbeitsort_plz}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{arbeitgeber}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{beruf}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{branche}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{arbeitszeit}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{eintrittsdatum}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{zeitarbeit}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}

\item{\code{corona}}{\link{JobSearchResponseFacettenInnerBefristung} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobSearchResponseFacettenInner-new}{\code{JobSearchResponseFacettenInner$new()}}
\item \href{#method-JobSearchResponseFacettenInner-toJSON}{\code{JobSearchResponseFacettenInner$toJSON()}}
\item \href{#method-JobSearchResponseFacettenInner-toList}{\code{JobSearchResponseFacettenInner$toList()}}
\item \href{#method-JobSearchResponseFacettenInner-toSimpleType}{\code{JobSearchResponseFacettenInner$toSimpleType()}}
\item \href{#method-JobSearchResponseFacettenInner-fromJSON}{\code{JobSearchResponseFacettenInner$fromJSON()}}
\item \href{#method-JobSearchResponseFacettenInner-toJSONString}{\code{JobSearchResponseFacettenInner$toJSONString()}}
\item \href{#method-JobSearchResponseFacettenInner-fromJSONString}{\code{JobSearchResponseFacettenInner$fromJSONString()}}
\item \href{#method-JobSearchResponseFacettenInner-validateJSON}{\code{JobSearchResponseFacettenInner$validateJSON()}}
\item \href{#method-JobSearchResponseFacettenInner-toString}{\code{JobSearchResponseFacettenInner$toString()}}
\item \href{#method-JobSearchResponseFacettenInner-isValid}{\code{JobSearchResponseFacettenInner$isValid()}}
\item \href{#method-JobSearchResponseFacettenInner-getInvalidFields}{\code{JobSearchResponseFacettenInner$getInvalidFields()}}
\item \href{#method-JobSearchResponseFacettenInner-print}{\code{JobSearchResponseFacettenInner$print()}}
\item \href{#method-JobSearchResponseFacettenInner-clone}{\code{JobSearchResponseFacettenInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobSearchResponseFacettenInner class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$new(
  befristung = NULL,
  behinderung = NULL,
  pav = NULL,
  berufsfeld = NULL,
  arbeitsort = NULL,
  ausbildungsart = NULL,
  veroeffentlichtseit = NULL,
  schulbildung = NULL,
  arbeitsort_plz = NULL,
  arbeitgeber = NULL,
  beruf = NULL,
  branche = NULL,
  arbeitszeit = NULL,
  eintrittsdatum = NULL,
  zeitarbeit = NULL,
  corona = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{befristung}}{befristung}

\item{\code{behinderung}}{behinderung}

\item{\code{pav}}{pav}

\item{\code{berufsfeld}}{berufsfeld}

\item{\code{arbeitsort}}{arbeitsort}

\item{\code{ausbildungsart}}{ausbildungsart}

\item{\code{veroeffentlichtseit}}{veroeffentlichtseit}

\item{\code{schulbildung}}{schulbildung}

\item{\code{arbeitsort_plz}}{arbeitsort_plz}

\item{\code{arbeitgeber}}{arbeitgeber}

\item{\code{beruf}}{beruf}

\item{\code{branche}}{branche}

\item{\code{arbeitszeit}}{arbeitszeit}

\item{\code{eintrittsdatum}}{eintrittsdatum}

\item{\code{zeitarbeit}}{zeitarbeit}

\item{\code{corona}}{corona}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobSearchResponseFacettenInner as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobSearchResponseFacettenInner (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobSearchResponseFacettenInner to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobSearchResponseFacettenInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobSearchResponseFacettenInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobSearchResponseFacettenInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobSearchResponseFacettenInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobSearchResponseFacettenInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobSearchResponseFacettenInner and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobSearchResponseFacettenInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
