% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_search_response_facetten_inner_befristung.R
\docType{class}
\name{JobSearchResponseFacettenInnerBefristung}
\alias{JobSearchResponseFacettenInnerBefristung}
\title{JobSearchResponseFacettenInnerBefristung}
\format{
An \code{R6Class} generator object
}
\description{
JobSearchResponseFacettenInnerBefristung Class

JobSearchResponseFacettenInnerBefristung Class
}
\details{
Create a new JobSearchResponseFacettenInnerBefristung
}
\examples{

## ------------------------------------------------
## Method `JobSearchResponseFacettenInnerBefristung$toList`
## ------------------------------------------------

# convert array of JobSearchResponseFacettenInnerBefristung (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{counts}}{object [optional]}

\item{\code{maxCount}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobSearchResponseFacettenInnerBefristung-new}{\code{JobSearchResponseFacettenInnerBefristung$new()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-toJSON}{\code{JobSearchResponseFacettenInnerBefristung$toJSON()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-toList}{\code{JobSearchResponseFacettenInnerBefristung$toList()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-toSimpleType}{\code{JobSearchResponseFacettenInnerBefristung$toSimpleType()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-fromJSON}{\code{JobSearchResponseFacettenInnerBefristung$fromJSON()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-toJSONString}{\code{JobSearchResponseFacettenInnerBefristung$toJSONString()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-fromJSONString}{\code{JobSearchResponseFacettenInnerBefristung$fromJSONString()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-validateJSON}{\code{JobSearchResponseFacettenInnerBefristung$validateJSON()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-toString}{\code{JobSearchResponseFacettenInnerBefristung$toString()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-isValid}{\code{JobSearchResponseFacettenInnerBefristung$isValid()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-getInvalidFields}{\code{JobSearchResponseFacettenInnerBefristung$getInvalidFields()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-print}{\code{JobSearchResponseFacettenInnerBefristung$print()}}
\item \href{#method-JobSearchResponseFacettenInnerBefristung-clone}{\code{JobSearchResponseFacettenInnerBefristung$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobSearchResponseFacettenInnerBefristung class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$new(
  counts = NULL,
  maxCount = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{counts}}{counts}

\item{\code{maxCount}}{maxCount}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobSearchResponseFacettenInnerBefristung as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobSearchResponseFacettenInnerBefristung (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobSearchResponseFacettenInnerBefristung to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobSearchResponseFacettenInnerBefristung
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobSearchResponseFacettenInnerBefristung
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobSearchResponseFacettenInnerBefristung in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobSearchResponseFacettenInnerBefristung
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobSearchResponseFacettenInnerBefristung
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobSearchResponseFacettenInnerBefristung and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobSearchResponseFacettenInnerBefristung
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseFacettenInnerBefristung-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseFacettenInnerBefristung-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseFacettenInnerBefristung$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
