% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/job_search_response_stellenangebote_inner_arbeitsort.R
\docType{class}
\name{JobSearchResponseStellenangeboteInnerArbeitsort}
\alias{JobSearchResponseStellenangeboteInnerArbeitsort}
\title{JobSearchResponseStellenangeboteInnerArbeitsort}
\format{
An \code{R6Class} generator object
}
\description{
JobSearchResponseStellenangeboteInnerArbeitsort Class

JobSearchResponseStellenangeboteInnerArbeitsort Class
}
\details{
Create a new JobSearchResponseStellenangeboteInnerArbeitsort
}
\examples{

## ------------------------------------------------
## Method `JobSearchResponseStellenangeboteInnerArbeitsort$toList`
## ------------------------------------------------

# convert array of JobSearchResponseStellenangeboteInnerArbeitsort (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{plz}}{integer [optional]}

\item{\code{ort}}{character [optional]}

\item{\code{strasse}}{character [optional]}

\item{\code{region}}{character [optional]}

\item{\code{land}}{character [optional]}

\item{\code{koordinaten}}{\link{JobSearchResponseStellenangeboteInnerArbeitsortKoordinaten} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-new}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$new()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-toJSON}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$toJSON()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-toList}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$toList()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-toSimpleType}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$toSimpleType()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-fromJSON}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$fromJSON()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-toJSONString}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$toJSONString()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-fromJSONString}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$fromJSONString()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-validateJSON}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$validateJSON()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-toString}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$toString()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-isValid}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$isValid()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-getInvalidFields}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$getInvalidFields()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-print}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$print()}}
\item \href{#method-JobSearchResponseStellenangeboteInnerArbeitsort-clone}{\code{JobSearchResponseStellenangeboteInnerArbeitsort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-new"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new JobSearchResponseStellenangeboteInnerArbeitsort class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$new(
  plz = NULL,
  ort = NULL,
  strasse = NULL,
  region = NULL,
  land = NULL,
  koordinaten = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plz}}{plz}

\item{\code{ort}}{ort}

\item{\code{strasse}}{strasse}

\item{\code{region}}{region}

\item{\code{land}}{land}

\item{\code{koordinaten}}{koordinaten}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert to an R object. This method is deprecated. Use `toSimpleType()` instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-toList"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-toList}{}}}
\subsection{Method \code{toList()}}{
Convert to a List

Convert the R6 object to a list to work more easily with other tooling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JobSearchResponseStellenangeboteInnerArbeitsort as a base R list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# convert array of JobSearchResponseStellenangeboteInnerArbeitsort (x) to a data frame
\dontrun{
library(purrr)
library(tibble)
df <- x |> map(\(y)y$toList()) |> map(as_tibble) |> list_rbind()
df
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-toSimpleType"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-toSimpleType}{}}}
\subsection{Method \code{toSimpleType()}}{
Convert JobSearchResponseStellenangeboteInnerArbeitsort to a base R type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$toSimpleType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A base R type, e.g. a list or numeric/character array.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of JobSearchResponseStellenangeboteInnerArbeitsort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobSearchResponseStellenangeboteInnerArbeitsort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$toJSONString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to `jsonlite::toJSON`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
JobSearchResponseStellenangeboteInnerArbeitsort in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of JobSearchResponseStellenangeboteInnerArbeitsort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of JobSearchResponseStellenangeboteInnerArbeitsort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to JobSearchResponseStellenangeboteInnerArbeitsort and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-toString"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of JobSearchResponseStellenangeboteInnerArbeitsort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-print"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JobSearchResponseStellenangeboteInnerArbeitsort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JobSearchResponseStellenangeboteInnerArbeitsort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JobSearchResponseStellenangeboteInnerArbeitsort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
